% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PandemicPrediction.R
\docType{package}
\name{PandemicPrediction}
\alias{PandemicPrediction-package}
\alias{PandemicPrediction}
\title{PandemicPrediction}
\description{
A study-package for PandemicPrediction study
}
\author{
\strong{Maintainer}: Egill Fridgeirsson \email{Fridgeirsson@ohdsi.org}

}
\keyword{internal}
